+++ Cesium3DTilesSelection/src/RasterOverlayTileProvider.cpp
@@ -8,7 +8,7 @@
 #include "Cesium3DTilesSelection/spdlog-cesium.h"
 
 #include <CesiumAsync/IAssetResponse.h>
-#include <CesiumGltf/GltfReader.h>
+#include <CesiumGltfReader/GltfReader.h>
 #include <CesiumUtility/Tracing.h>
 #include <CesiumUtility/joinToString.h>
 
@@ -16,11 +16,13 @@ using namespace CesiumAsync;
 using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
 using namespace CesiumGltf;
+using namespace CesiumGltfReader;
 using namespace CesiumUtility;
 
 namespace Cesium3DTilesSelection {
 
-/*static*/ CesiumGltf::GltfReader RasterOverlayTileProvider::_gltfReader{};
+/*static*/ CesiumGltfReader::GltfReader
+    RasterOverlayTileProvider::_gltfReader{};
 
 RasterOverlayTileProvider::RasterOverlayTileProvider(
     RasterOverlay& owner,
@@ -175,7 +177,7 @@ RasterOverlayTileProvider::loadTileImageFromUrl(
 
             const gsl::span<const std::byte> data = pResponse->data();
 
-            CesiumGltf::ImageReaderResult loadedImage =
+            CesiumGltfReader::ImageReaderResult loadedImage =
                 RasterOverlayTileProvider::_gltfReader.readImage(data);
 
             if (!loadedImage.errors.empty()) {