+++ CesiumGltfReader/src/decodeDataUrls.h
@@ -1,6 +1,6 @@
 #pragma once
 
-namespace CesiumGltf {
+namespace CesiumGltfReader {
 
 struct ModelReaderResult;
 class GltfReader;
@@ -9,4 +9,4 @@ void decodeDataUrls(
     const GltfReader& reader,
     ModelReaderResult& readModel,
     bool clearDecodedDataUrls);
-} // namespace CesiumGltf
+} // namespace CesiumGltfReader