+++ CesiumGltfWriter/src/ExtensionWriter.h
@@ -5,7 +5,7 @@
 #include <any>
 #include <unordered_map>
 
-namespace CesiumGltf {
+namespace CesiumGltfWriter {
 void writeExtensions(
     const std::unordered_map<std::string, std::any>& extensions,
     CesiumJsonWriter::JsonWriter& jsonWriter);