+++ tools/generate-classes/generate.js
@@ -1,21 +1,24 @@
 const fs = require("fs");
-const getNameFromSchema = require("./getNameFromSchema");
+const getNameFromTitle = require("./getNameFromTitle");
 const indent = require("./indent");
 const lodash = require("lodash");
 const path = require("path");
 const resolveProperty = require("./resolveProperty");
 const unindent = require("./unindent");
 
-function generate(options, schema) {
+function generate(options, schema, writers) {
   const {
     schemaCache,
     outputDir,
     readerOutputDir,
     config,
     namespace,
+    readerNamespace,
+    writerNamespace,
+    extensions,
   } = options;
 
-  const name = getNameFromSchema(config, schema);
+  const name = getNameFromTitle(config, schema.title);
   const thisConfig = config.classes[schema.title] || {};
 
   console.log(`Generating ${name}`);
@@ -25,7 +28,7 @@ function generate(options, schema) {
   let base = "CesiumUtility::ExtensibleObject";
   if (schema.allOf && schema.allOf.length > 0 && schema.allOf[0].$ref) {
     const baseSchema = schemaCache.load(schema.allOf[0].$ref);
-    base = getNameFromSchema(config, baseSchema);
+    base = getNameFromTitle(config, baseSchema.title);
   }
 
   const required = schema.required || [];
@@ -39,7 +42,9 @@ function generate(options, schema) {
         key,
         schema.properties[key],
         required,
-        namespace
+        namespace,
+        readerNamespace,
+        writerNamespace
       )
     )
     .filter((property) => property !== undefined);
@@ -51,14 +56,14 @@ function generate(options, schema) {
   schemaCache.popContext();
 
   let headers = lodash.uniq([
-    `"Library.h"`,
-    getIncludeFromName(base),
+    `"${namespace}/Library.h"`,
+    getIncludeFromName(base, namespace),
     ...lodash.flatten(properties.map((property) => property.headers)),
   ]);
 
   // Prevent header from including itself for recursive types like Tile
   headers = headers.filter((header) => {
-    return header !== `"${name}.h"`;
+    return header !== `"${namespace}/${name}.h"`;
   });
 
   headers.sort();
@@ -93,7 +98,12 @@ function generate(options, schema) {
         }
     `;
 
-  const headerOutputDir = path.join(outputDir, "include", namespace);
+  const headerOutputDir = path.join(
+    outputDir,
+    "generated",
+    "include",
+    namespace
+  );
   fs.mkdirSync(headerOutputDir, { recursive: true });
   const headerOutputPath = path.join(
     headerOutputDir,
@@ -102,7 +112,7 @@ function generate(options, schema) {
   fs.writeFileSync(headerOutputPath, unindent(header), "utf-8");
 
   let readerHeaders = lodash.uniq([
-    getReaderIncludeFromName(base),
+    getReaderIncludeFromName(base, readerNamespace),
     `<${namespace}/${name}.h>`,
     ...lodash.flatten(properties.map((property) => property.readerHeaders)),
   ]);
@@ -118,7 +128,7 @@ function generate(options, schema) {
     lodash.flatten(properties.map((property) => property.readerLocalTypes))
   );
 
-  const baseReader = getReaderName(base);
+  const baseReader = getReaderName(base, readerNamespace);
 
   // prettier-ignore
   const readerHeader = `
@@ -132,15 +142,15 @@ function generate(options, schema) {
           class ExtensionReaderContext;
         }
 
-        namespace ${namespace} {
+        namespace ${readerNamespace} {
           class ${name}JsonHandler : public ${baseReader}${thisConfig.extensionName ? `, public CesiumJsonReader::IExtensionJsonHandler` : ""} {
           public:
-            using ValueType = ${name};
+            using ValueType = ${namespace}::${name};
 
             ${thisConfig.extensionName ? `static inline constexpr const char* ExtensionName = "${thisConfig.extensionName}";` : ""}
 
             ${name}JsonHandler(const CesiumJsonReader::ExtensionReaderContext& context) noexcept;
-            void reset(IJsonHandler* pParentHandler, ${name}* pObject);
+            void reset(IJsonHandler* pParentHandler, ${namespace}::${name}* pObject);
 
             virtual IJsonHandler* readObjectKey(const std::string_view& str) override;
 
@@ -189,12 +199,11 @@ function generate(options, schema) {
             ` : ""}
 
           protected:
-            IJsonHandler* readObjectKey${name}(const std::string& objectType, const std::string_view& str, ${name}& o);
+            IJsonHandler* readObjectKey${name}(const std::string& objectType, const std::string_view& str, ${namespace}::${name}& o);
 
           private:
             ${indent(readerLocalTypes.join("\n\n"), 12)}
-
-            ${name}* _pObject = nullptr;
+            ${namespace}::${name}* _pObject = nullptr;
             ${indent(
               properties
                 .map((property) => formatReaderProperty(property))
@@ -205,8 +214,9 @@ function generate(options, schema) {
         }
   `;
 
-  const readerHeaderOutputDir = path.join(readerOutputDir, "generated");
+  const readerHeaderOutputDir = path.join(readerOutputDir, "generated", "src");
   fs.mkdirSync(readerHeaderOutputDir, { recursive: true });
+
   const readerHeaderOutputPath = path.join(
     readerHeaderOutputDir,
     name + "JsonHandler.h"
@@ -226,7 +236,10 @@ function generate(options, schema) {
     const initializerList = properties
       .filter((p) => p.readerType.toLowerCase().indexOf("jsonhandler") != -1)
       .map(
-        (p) => `_${p.cppSafeName}(${p.schemas && p.schemas.length > 0 ? varName : ""})`
+        (p) =>
+          `_${p.cppSafeName}(${
+            p.schemas && p.schemas.length > 0 ? varName : ""
+          })`
       )
       .join(", ");
     return initializerList == "" ? "" : ", " + initializerList;
@@ -236,51 +249,153 @@ function generate(options, schema) {
         // This file was generated by generate-classes.
         // DO NOT EDIT THIS FILE!
         #include "${name}JsonHandler.h"
+
         #include <${namespace}/${name}.h>
         ${readerHeadersImpl.map((header) => `#include ${header}`).join("\n")}
         #include <cassert>
         #include <string>
 
-        using namespace ${namespace};
+        namespace ${readerNamespace} {
 
         ${name}JsonHandler::${name}JsonHandler(const CesiumJsonReader::ExtensionReaderContext& context) noexcept : ${baseReader}(context)${generateReaderOptionsInitializerList(properties, 'context')} {}
 
-        void ${name}JsonHandler::reset(CesiumJsonReader::IJsonHandler* pParentHandler, ${name}* pObject) {
+        void ${name}JsonHandler::reset(CesiumJsonReader::IJsonHandler* pParentHandler, ${namespace}::${name}* pObject) {
           ${baseReader}::reset(pParentHandler, pObject);
           this->_pObject = pObject;
         }
 
         CesiumJsonReader::IJsonHandler* ${name}JsonHandler::readObjectKey(const std::string_view& str) {
           assert(this->_pObject);
-          return this->readObjectKey${name}(${name}::TypeName, str, *this->_pObject);
+          return this->readObjectKey${name}(${namespace}::${name}::TypeName, str, *this->_pObject);
         }
 
         ${thisConfig.extensionName ? `
         void ${name}JsonHandler::reset(CesiumJsonReader::IJsonHandler* pParentHandler, CesiumUtility::ExtensibleObject& o, const std::string_view& extensionName) {
           std::any& value =
-              o.extensions.emplace(extensionName, ${name}())
+              o.extensions.emplace(extensionName, ${namespace}::${name}())
                   .first->second;
           this->reset(
               pParentHandler,
-              &std::any_cast<${name}&>(value));
+              &std::any_cast<${namespace}::${name}&>(value));
         }
         ` : ""}
 
-        CesiumJsonReader::IJsonHandler* ${name}JsonHandler::readObjectKey${name}(const std::string& objectType, const std::string_view& str, ${name}& o) {
+        CesiumJsonReader::IJsonHandler* ${name}JsonHandler::readObjectKey${name}(const std::string& objectType, const std::string_view& str, ${namespace}::${name}& o) {
           using namespace std::string_literals;
 
+          ${properties.length > 0 ? `
           ${indent(
             properties
               .map((property) => formatReaderPropertyImpl(property))
               .join("\n"),
             10
-          )}
+          )}` : `(void)o;`}
 
           return this->readObjectKey${removeNamespace(base)}(objectType, str, *this->_pObject);
         }
 
         ${indent(readerLocalTypesImpl.join("\n\n"), 8)}
+        } // namespace ${readerNamespace}
   `;
+
+  const writeForwardDeclaration = `struct ${name};`;
+
+  const writeInclude = `#include <${namespace}/${name}.h>`;
+
+  const writeDeclaration = `
+        struct ${name}JsonWriter {
+          using ValueType = ${namespace}::${name};
+
+          ${
+            thisConfig.extensionName
+              ? `static inline constexpr const char* ExtensionName = "${thisConfig.extensionName}";`
+              : ""
+          }
+
+          static void write(
+              const ${namespace}::${name}& obj,
+              CesiumJsonWriter::JsonWriter& jsonWriter,
+              const CesiumJsonWriter::ExtensionWriterContext& context);
+        };
+  `;
+
+  const writeJsonDeclaration = `
+        void writeJson(
+            const ${namespace}::${name}& obj,
+            CesiumJsonWriter::JsonWriter& jsonWriter,
+            const CesiumJsonWriter::ExtensionWriterContext& context);
+  `;
+
+  const writeDefinition = `
+        void ${name}JsonWriter::write(
+            const ${namespace}::${name}& obj,
+            CesiumJsonWriter::JsonWriter& jsonWriter,
+            const CesiumJsonWriter::ExtensionWriterContext& context) {
+          writeJson(obj, jsonWriter, context);
+        }
+  `;
+
+  const writeJsonDefinition = `
+        void writeJson(
+            const ${namespace}::${name}& obj,
+            CesiumJsonWriter::JsonWriter& jsonWriter,
+            const CesiumJsonWriter::ExtensionWriterContext& context) {
+          jsonWriter.StartObject();
+
+          ${indent(
+            properties
+              .map((property) => formatWriterPropertyImpl(property))
+              .join("\n\n"),
+            10
+          )}
+
+          ${
+            schema.properties.extensions
+              ? `
+            if (!obj.extensions.empty()) {
+              jsonWriter.Key("extensions");
+              writeJsonExtensions(obj, jsonWriter, context);
+            }
+          `
+              : ""
+          }
+
+          ${
+            schema.properties.extras
+              ? `
+            if (!obj.extras.empty()) {
+              jsonWriter.Key("extras");
+              writeJson(obj.extras, jsonWriter, context);
+            }
+          `
+              : ""
+          }
+
+          jsonWriter.EndObject();
+        }
+  `;
+
+  const writeExtensionsRegistration = `
+        ${
+          extensions[schema.title]
+            ? extensions[schema.title]
+                .map((extension) => {
+                  return `context.registerExtension<${namespace}::${name}, ${extension.className}JsonWriter>();`;
+                })
+                .join("\n")
+            : ""
+        }
+  `;
+
+  writers.push({
+    writeInclude,
+    writeForwardDeclaration,
+    writeDeclaration,
+    writeJsonDeclaration,
+    writeDefinition,
+    writeJsonDefinition,
+    writeExtensionsRegistration,
+  });
 
   if (options.oneHandlerFile) {
     const readerSourceOutputPath = path.join(
@@ -336,6 +451,60 @@ function formatReaderPropertyImpl(property) {
   return `if ("${property.name}"s == str) return property("${property.name}", this->_${property.cppSafeName}, o.${property.cppSafeName});`;
 }
 
+function formatWriterPropertyImpl(property) {
+  let result = "";
+
+  const type = property.type;
+  const defaultValue = property.defaultValueWriter || property.defaultValue;
+
+  const isId = property.requiredId !== undefined;
+  const isOptionalId = property.requiredId === false;
+  const isRequiredEnum = property.requiredEnum === true;
+  const isVector = type.startsWith("std::vector");
+  const isMap = type.startsWith("std::unordered_map");
+  const isOptional = type.startsWith("std::optional");
+
+  // Somewhat opinionated but it's helpful to see byteOffset: 0 in accessors and bufferViews
+  const requiredPropertyOverride = ["byteOffset"];
+
+  const hasDefaultValueGuard =
+    !isId &&
+    !isRequiredEnum &&
+    !requiredPropertyOverride.includes(property.cppSafeName) &&
+    defaultValue !== undefined;
+  const hasDefaultVectorGuard = hasDefaultValueGuard && isVector;
+  const hasEmptyGuard = isVector || isMap;
+  const hasOptionalGuard = isOptional;
+  const hasNegativeIndexGuard = isOptionalId;
+  const hasGuard =
+    hasDefaultValueGuard ||
+    hasEmptyGuard ||
+    hasOptionalGuard ||
+    hasNegativeIndexGuard;
+
+  if (hasDefaultVectorGuard) {
+    result += `static const ${type} ${property.cppSafeName}Default = ${defaultValue};\n`;
+    result += `if (obj.${property.cppSafeName} != ${property.cppSafeName}Default) {\n`;
+  } else if (hasDefaultValueGuard) {
+    result += `if (obj.${property.cppSafeName} != ${defaultValue}) {\n`;
+  } else if (hasEmptyGuard) {
+    result += `if (!obj.${property.cppSafeName}.empty()) {\n`;
+  } else if (hasNegativeIndexGuard) {
+    result += `if (obj.${property.cppSafeName} > -1) {\n`;
+  } else if (hasOptionalGuard) {
+    result += `if (obj.${property.cppSafeName}.has_value()) {\n`;
+  }
+
+  result += `jsonWriter.Key("${property.name}");\n`;
+  result += `writeJson(obj.${property.cppSafeName}, jsonWriter, context);\n`;
+
+  if (hasGuard) {
+    result += "}\n";
+  }
+
+  return result;
+}
+
 function privateSpecConstructor(name) {
   return `
     private:
@@ -349,39 +518,50 @@ function privateSpecConstructor(name) {
 
 const qualifiedTypeNameRegex = /(?:(?<namespace>.+)::)?(?<name>.+)/;
 
-function getIncludeFromName(name) {
+function getReaderNamespace(namespace, readerNamespace) {
+  if (namespace === "CesiumUtility") {
+    return "CesiumJsonReader";
+  }
+  return readerNamespace;
+}
+
+function getIncludeFromName(name, namespace) {
   const pieces = name.match(qualifiedTypeNameRegex);
   if (pieces && pieces.groups && pieces.groups.namespace) {
-    return `<${pieces.groups.namespace}/${pieces.groups.name}.h>`;
+    if (pieces.groups.namespace === namespace) {
+      return `"${namespace}/${pieces.groups.name}.h"`;
+    } else {
+      return `<${pieces.groups.namespace}/${pieces.groups.name}.h>`;
+    }
   } else {
-    return `"${name}.h"`;
+    return `"${namespace}/${name}.h"`;
   }
 }
 
-function getReaderIncludeFromName(name) {
+function getReaderIncludeFromName(name, readerNamespace) {
   const pieces = name.match(qualifiedTypeNameRegex);
   if (pieces && pieces.groups && pieces.groups.namespace) {
-    let namespace = pieces.groups.namespace;
-
-    // CesiumUtility types have their corresponding reader classes in CesiumJsonReader
-    if (namespace === "CesiumUtility") {
-      namespace = "CesiumJsonReader";
+    const namespace = getReaderNamespace(
+      pieces.groups.namespace,
+      readerNamespace
+    );
+    if (namespace === readerNamespace) {
+      return `"${pieces.groups.name}JsonHandler.h"`;
+    } else {
+      return `<${namespace}/${pieces.groups.name}JsonHandler.h>`;
     }
-    return `<${namespace}/${pieces.groups.name}JsonHandler.h>`;
   } else {
     return `"${name}JsonHandler.h"`;
   }
 }
 
-function getReaderName(name) {
+function getReaderName(name, readerNamespace) {
   const pieces = name.match(qualifiedTypeNameRegex);
   if (pieces && pieces.groups && pieces.groups.namespace) {
-    let namespace = pieces.groups.namespace;
-
-    // CesiumUtility types have their corresponding reader classes in CesiumJsonReader
-    if (namespace === "CesiumUtility") {
-      namespace = "CesiumJsonReader";
-    }
+    const namespace = getReaderNamespace(
+      pieces.groups.namespace,
+      readerNamespace
+    );
     return `${namespace}::${pieces.groups.name}JsonHandler`;
   } else {
     return `${name}JsonHandler`;