+++ CHANGES.md
@@ -1,5 +1,21 @@
 # Change Log
 
+### v0.11.0 - 2022-01-03
+
+##### Breaking Changes :mega:
+
+- The `CesiumGltfReader` project now uses the `CesiumGltfReader` namespace instead of the `CesiumGltf` namespace.
+- The `CesiumGltfWriter` project now uses the `CesiumGltfWriter` namespace instead of the `CesiumGltf` namespace.
+- The `Cesium3DTilesReader` project now uses the `Cesium3DTilesReader` namespace instead of the `Cesium3DTiles` namespace.
+
+##### Additions :tada:
+
+- Added `Cesium3DTilesWriter` library.
+
+##### Fixes :wrench:
+
+- Fixed a bug in `QuadtreeRasterOverlayTileProvider` that caused incorrect level-of-detail selection for overlays that use a global (or otherwise large) tiling scheme but have non-global (or otherwise smaller) coverage.
+
 ### v0.10.0 - 2021-12-01
 
 ##### Breaking Changes :mega: